Moved functions to functions.py
