+++ generateData/utils/functions.py
@@ -1,6 +1,8 @@
 
 import numpy as np
 import cv2
+from utils.controller import Controller # Original controller implementation by Rafael Sonderegger, opted using simplePid library instead
+from simple_pid import PID # Simple PID library by Brett Beauregard
 
 def findEdges(image):
     hsv = cv2.cvtColor(image, cv2.COLOR_BGR2HSV)
@@ -74,24 +76,108 @@ def calculateTargetPoint(image, widthOfTrack, freq, scale_dist, Amplitude, t):
 
     return targetPoint, estimatedMiddlePoint, vector_track_normalized ,vector_track_perp_normalized
 
-def controllerSteeringAngle(angle, prevSteeringAngle):
-    #PID control of steering angle
-    Kp = 0.5
-    Ki = 0.1    
-    Kd = -0.2
-    steeringAngle = Kp * angle + Kd * (angle - prevSteeringAngle)
-    return steeringAngle
 
-def controllerSpeed(targetSpeed, currentSpeed, prevSpeedError):
-    Kp = 0.5
-    Ki = 0.1
-    Kd = 0.1
-    error = targetSpeed - currentSpeed
-    acceleration = Kp * error + Kd * (error - prevSpeedError)
+def find_edge_1dStrip(array, direction):
+    # Find edge point of a 1D array. 
+    # If none is found return -1
+    starting_point = int(len(array) // 2)
+    idx = -1
+    if direction == 'left':
+        for i in range(starting_point, -1, -1):
+            if array[i] != 0:
+                idx = i
+                break
+    elif direction == 'right':
+        for i in range(starting_point, len(array)):
+            if array[i] != 0:
+                idx = i
+                break
+    return idx
+
+def find_middle_point(strip_1d):
+    # Check if there is edge point for both left and right side of track.
+    # If none is found set border of strip as edge point
+    idx1 = find_edge_1dStrip(strip_1d, 'left')
+    idx2 = find_edge_1dStrip(strip_1d, 'right')
+
+    if idx1 == -1:
+        idx1 = 0
+    if idx2 == -1:
+        idx2 = len(strip_1d) - 1
+
+    idx_middle = int((idx1 + idx2) / 2)
+    return idx_middle
+
+def calculateDistAngle(idx_middle_upper, idx_middle_lower, strip_width, strip_height):
+    # Calculate distance and angle from middle of the track
+    # idx_middle_upper: index of middle point on upper edge of strip
+    # idx_middle_lower: index of middle point on lower edge of strip
+    # strip_width: width of the strip
+    # strip_height: height of the strip
+    # return: distance and angle
+    
+    # Compute distance to middleline
+    distance_to_middleline = strip_width // 2 - idx_middle_lower
+    # Compute angular error
+    upper_lenght_to_target = strip_width // 2 - idx_middle_upper
+    angle_to_target = np.arctan(upper_lenght_to_target / strip_height)
+    return distance_to_middleline, angle_to_target
+
+
+def processImage(image):
+    # Cropping image down to a strip
+    strip_height = 20
+    strip_width = 96
+    middle_height = 65
+    top = int(middle_height - strip_height / 2)
+    bottom = int(middle_height + strip_height / 2)
+    # Crop the strip from the image
+    strip = image[top:bottom, :]
+
+    ## Mask where only edge is retained
+    hsv = cv2.cvtColor(strip, cv2.COLOR_BGR2HSV)
+    mask_green = cv2.inRange(hsv, (36, 25, 25), (70, 255,255))
+    imask_green = mask_green>0
+    gray_mask = imask_green.astype(np.uint8)
+    gray_mask = gray_mask*255
+    # Only use two edges of the strip: Upper and lower and find edge points coordinates
+    upper_edge = gray_mask[0, :]
+    lower_edge = gray_mask[strip_height - 1, :]
+    # Get index of middle point on the upper and lower edge
+    idx_middle_upper = find_middle_point(upper_edge)
+    idx_middle_lower = find_middle_point(lower_edge)
+
+    distance, angle = calculateDistAngle(idx_middle_upper, idx_middle_lower, strip_width, strip_height)
+    return distance, angle
+
+
+def calculateAction(observation , target_velocity):
+
+    # Initialize controllers
+    pid_angle = PID(0.5, -0.01, 0.05, setpoint=0)
+    pid_distance = PID(0.5, -0.005, 0.05, setpoint=0)
+    pid_velocity = PID(0.05, -0.1, 0.2, setpoint=target_velocity)
+    
+
+    # Distinguish observation type
+    image = observation['image']
+    velocity = observation['velocity']
+
+    # Get distance from processed image
+    error_dist, error_ang = processImage(image)
+
+    # Get control outputs from PD controllers
+    control_ang = pid_angle(error_ang)
+    control_dist = pid_distance(error_dist)
+    control_vel = pid_velocity(velocity)
+
+    print("Control outputs: ", control_ang, control_dist, control_vel)
+    acc = control_vel
     breaking = 0
-    prevSpeedError = error
-
-    if acceleration < 0 :
-        breaking = -acceleration
-        acceleration = 0 
-    return acceleration, breaking
+    if acc < 0:
+        acc = 0
+        breaking = -control_vel
+    
+    # Calculate and return final action
+    action = (control_ang , acc, breaking)
+    return action
